backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
requests==2.31.0
cryptography==42.0.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.5.3
pymongo==4.6.1
dnspython==2.4.2